add failsafe to number input
added failsafe to number input.
changed assign project person to a bool so we can output success message when its done.
added MoveCursorBottom to move the cursor to the bottom after selecting an index to avoid writing the results under the selcted index instead it will now be after all menu options